+++ packages/@lou.codes/iterables/src/asynchronous/repeat.ts
@@ -20,8 +20,8 @@ export const repeat =
 		createIterableIterator(async function* () {
 			// eslint-disable-next-line functional/no-conditional-statements
 			if (times === Infinity) {
-				// eslint-disable-next-line functional/no-loop-statements, @typescript-eslint/no-unnecessary-condition
-				while (true) {
+				// eslint-disable-next-line functional/no-loop-statements
+				for (;;) {
 					yield item;
 				}
 				// eslint-disable-next-line functional/no-conditional-statements